study_buddy/requirements.txt
python-dotenv
requests
PyPDF2